TT2WB util=50% ep=4 addrLen=84 maxPacket=40 4x2 DataFSM hasError
UsbDataRxFsm hasError when FSM onEntry initializes crcError=false
This has now been inverted so it starts as crcError=true and a
correct CRC16 at the end the only way it can make a transition
to crcError=false.

This impacts hasError which now only indicates no-error
(hasError=false) when the FSM successfully completes all states
with the CRC16 being correct at the end.