UsbDeviceTop.v: timer width reductions
Tested with FS and LS.

-  reg        [6:0]    timer_counter;
+  reg        [5:0]    timer_counter;
-  reg        [8:0]    rx_packet_errorTimeout_counter;
+  reg        [6:0]    rx_packet_errorTimeout_counter;
-  reg        [22:0]   rx_timerLong_counter;
+  reg        [20:0]   rx_timerLong_counter;

In particular rx_timerLong_counter uses MSB as overflow detection,
so this is one bit longer than any count.  assertions added to
project generation to confirm this is always maintained.

This should result in a marginal gain in resource usage.

The other counter width changes are a reduction based on the examining
how they are used and removing what looks like overprovisioning width
to count is possible.