UsbDeviceTop.v: make UsbTimer.errorTimer quiescent until use
This timer appears to only be used when a stuffingError occurs,
The purpose appears to be ignore data while state changes on D+/D-
occur, when the lines go stable state for 20 bit times then reset
back to IDLE.

Since a stuffingError should be a rare event, we can at least
save power.

Not sure why we can't just wait for SYNC (will check USB spec) and
remove this timer completely, will investigate.