TT2WB: Hold out8 state, improve coverage
Defensive change, maybe not needed.
This ensures the out8 is held for multiple TT interface cycles, just
incase it is not possible to read and write in the same cycle, or not
possible to pipeline reliably.

Maybe mux bypass is not needed either as the pipeline speedup is
redundant due to limitations.

Test DI3 and DO3 to improve coverage.